Restructure ruby files into lib directory
