+++ lib/cli.rb
