+++ lib/mission.rb
