+++ lib/mission_control.rb
