+++ lib/mission_plan.rb
