+++ lib/mission_reporter.rb
