+++ lib/space_craft.rb
